Add word counter using the new mutable trie
Modified wc.hs to use the mutable trie type instead of hash table
as the cunting lexicon. Compiles and tests OK.
However, speed falls short of expectations: about the same as the
word counter using the immutable trie, and only 50% of the speed
of the hash table version. Amended with new .gitignore.