Add Functor, Monad and Applicative instances
Continuing development of the immutable Trie module. Found a way to make
Trie into Monad so that the correspnding fmap is "what it should be".
Applicative seems to work "as expected" too.
Definition of (>>=) probably inefficient, proceeds through toList and
fromList => replace with "direct" definitions.
Added also skeleton for Traversable instance, and boilerplate Foldable
instance defined in terms of Traversable.